requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
httpx
beautifulsoup4
pandas
sqlalchemy
psycopg[binary]
redis
pgvector
mcp
pypdf
langchain
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
langgraph
yfinance
akshare
pytest
